sender_rp2.py
Removed the clock from the equation. The Python libraries proved to work differently on different devices (small differences in responses from the time library; on a single device, the library works perfectly, but when comparing it with the one implemented on a second device, I observed minor, unilinear differences). This made syncing two independent clocks unnecessarily complicated for the scope of the project. I concluded that if I were to write my own clock based on the time library, I might as well trust the accuracy of the time.sleep() function. The solution is not ideal, and frankly, it's disappointingly simple, but as mentioned before, there is no point in creating a clock structure if I want to use Python and its libraries. If in the future I decide that implementing a clock is necessary (probably when trying to speed up data transfer), I will rewrite the code in C or C++.